+++ source/gnss/gnssReceiver.cpp
@@ -134,7 +134,7 @@ void GnssReceiver::deleteObservation(UInt idTrans, UInt idEpoch)
 
 /***********************************************/
 
-void GnssReceiver::signalComposition(UInt /*idEpoch*/, const std::vector<GnssType> &types, std::vector<GnssType> &typesTrans, Matrix &A) const
+void GnssReceiver::signalCompositionDefault(const std::vector<GnssType> &types, std::vector<GnssType> &typesTrans, Matrix &A)
 {
   try
   {
@@ -218,6 +218,13 @@ void GnssReceiver::signalComposition(UInt /*idEpoch*/, const std::vector<GnssTyp
   {
     GROOPS_RETHROW(e)
   }
+}
+
+/***********************************************/
+
+void GnssReceiver::signalComposition(UInt /*idEpoch*/, const std::vector<GnssType> &types, std::vector<GnssType> &typesTrans, Matrix &A) const
+{
+  signalCompositionDefault(types, typesTrans, A);
 }
 
 /***********************************************/