+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Other:            ModelEquilibriumTide: Taking into account spatially varying density.
 - Other:            GnssTroposphere2TropoSinex: Revised interface.
 - Other:            GnssAntex2AntennaDefinition: Handling of changing antennas of a SVN.
+- Other:            GnssSinexBias2SignalBias: Transform combined X signals to actually transmitted signals.
 
 # Release 2025-11-15
 - New program:      StarCamera2Orbex (replaces GnssAttitude2Orbex).